requirements.txt
alembic==1.9.1
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
packaging==22.0
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.0
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.7
SQLAlchemy==1.4.46
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
